requirements.txt
# WeatherCore — Python Requirements
# Install: pip install -r requirements.txt

flask==3.1.3
flask-cors==6.0.0
mysql-connector-python==9.3.0
apscheduler==3.10.4
requests==2.33.0
python-dotenv==1.2.2
pandas==3.0.2

# Optional — Redis caching (Phase 5)
# redis==5.0.8

# Optional — ML prediction (Phase 5)
# scikit-learn==1.5.0
# numpy==2.0.0